requirements.txt
termcolor
python-cern-sso-krb
